+++ CHANGELOG.md
@@ -15,6 +15,9 @@ when set to `true`, if the lease container exists, connector will now process th
 #### Key Bug Fixes
 * Fixed issue in `CosmosDBSourceConnector` where no record being read when configured `connect.cosmos.offset.useLatest` to be false. [PR 516](https://github.com/microsoft/kafka-connect-cosmosdb/pull/516)
 
+#### Other Changes
+* Return from `poll` method once all records have been processed from the changeFeedProcessor batches. [PR 517](https://github.com/microsoft/kafka-connect-cosmosdb/pull/517)
+
 ### 1.8.0 (2023-04-12)
 #### New Features
 * Updated `azure-cosmos` version to 4.42.0.